+++ extensions/wayback-machine/src/lib.ts
@@ -1,5 +1,8 @@
 import { showToast, Toast } from "@raycast/api";
-import fetch from "cross-fetch";
+import { showFailureToast } from "@raycast/utils";
+
+export const WAYBACK_BASE_URL = "https://web.archive.org";
+export const WAYBACK_API_URL = "https://archive.org/wayback/available";
 
 export const urlRegex =
   /(http(s)?:\/\/.)?(www\.)?[-a-zA-Z0-9@:%._+~#=]{2,256}\.[a-z]{2,6}\b([-a-zA-Z0-9@:%_+.~#?&//=]*)/;
@@ -8,20 +11,17 @@ export async function savePage(webpageUrl: string) {
   const toast = await showToast({ style: Toast.Style.Animated, title: "Saving to Wayback Machine" });
 
   try {
-    const res = await fetch(`https://web.archive.org/save/${webpageUrl}`);
+    const res = await fetch(`${WAYBACK_BASE_URL}/save/${webpageUrl}`);
 
     if (res.status >= 400) {
-      toast.style = Toast.Style.Failure;
-      toast.title = "Failed to save to Wayback Machine";
+      await showFailureToast("Failed to save to Wayback Machine");
       return;
     }
 
     toast.style = Toast.Style.Success;
     toast.title = "Saved to Wayback Machine";
   } catch (err) {
-    console.log(err);
-    toast.style = Toast.Style.Failure;
-    toast.title = "Failed to save to Wayback Machine";
+    await showFailureToast(err, { title: "Failed to save to Wayback Machine" });
   }
 }